Allow Creating XMPP Conversations for XMPP Contacts
Pretty easy, actually. There was already code to list what inboxes you
can see based on what information we have about the customer.

So I just had to hook into that to say "XMPP" channels can be seen if
they have a JID.

>From there there's also a kind of default in the inboxes builder that if
we don't specify a source ID (in this case a JID) for the contact, it
can pull one out itself, so I built that too.

I don't think this code path actually uses it, because we actually
include the source ID in the listing, but it seemed like a good thing to
add.