Create Contact if Deeplink Doesn't Find Them
The assumption here is that some external system knows this person
exists, so if we are coming here it's probably because we're trying to
do something with them, like reach out to them in some way.

So make a new contact with the contact info we know about them so they
can move on. If it turns out later that we already had some information
on them, we can always merge later.