Create New XMPP Thread with Type and ThreadID
On inbound threads we copy the threadID and type of the initial message
we received, but when we are the initiator we need to pick something to
use as defaults, and preferably something where the user's response will
end up being sorted into the same conversation later.

But, not all clients support threading...

So we do a kind of heuristic here where we start the first conversation
as the threadless conversation because that will work on all clients.

But if there already is a conversation, and you've clicked the "New
Conversation" button anyway, then we assume you know what you're doing
and we make up a new threadID. Any client that supports threads will
reflect this back to us and their response will end up in the right
spot. Any client that doesn't will still have the response end up in the
threadless conversation, but there isn't really a lot we can do about
that.

We tried our best to do what was asked of us, and in the end what more
to life is there...